+++ bst-java/src/main/java/utybo/branchingstorytree/api/story/TagHolder.java
@@ -24,6 +24,11 @@ public class TagHolder
     {
         return tags.get(tag);
     }
+    
+    public String getTagOrDefault(final String tag, final String defaultVal)
+    {
+        return tags.getOrDefault(tag, defaultVal);
+    }
 
     public void putTag(final String tag, final String value)
     {